Stop overriding `iskeyword`
The idea behind including `.` in `iskeyword` was to enable completion
for the modules (e.g. `ansible.builtin.copy`) but not having other
operations like `cw` work as usual is confusing.

With completion solutions like `vim.lsp.completion` the `iskeyword`
addition is also no longer as important, because it prioritizes the word
boundary reported by the language server.